Add sandbox/_invoke-lambda
This is a fully general way of calling a lambda in a subprocesses, intended for use by the HTTP server
as well as events/queues.

It supports node & python runtimes. Subprocesses receive the lambda event on stdin, then respond by writing a JSON response to file-descriptor 3. This allows stdout/stderr to be written to normally.